.github/workflows: add main.yml for building fcode-utils upon push
This is a GitHub push action that generates a container image from
docker/Dockerfile using the fcode-utils-builder image and pushes the result
to ghcr.io/openbios/fcode-utils:master for public use. It is essentially
a standard Debian 11.2 image with the binaries copied into /usr/bin.

Note that this action also creates a zip artifact containing the binaries and
localvariables directories to allow simple build testing and verification
upon merge.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>