all: update from using internal bool implementation to stdbool
Since the bool type was included in the C99 standard nearly 30 years ago, it
is fairly safe to say that all modern compilers contain support for stdbool.

Remove the internal bool implementation from shared/types.h and then switch
over to using stdbool instead. The changes are fairly simple: change TRUE
and FALSE to lower case, and then remove the BOOLVAL() and INVERSE() macros
which are now no longer required.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>